moved error bag interfaces inside form component
